Clean up frontend codebase: remove unused files and consolidate contexts
This cleanup removes 6,426 lines of code across 35 files, improving code
maintainability and reducing bundle size:

**Context Files (High Priority)**
- Consolidated duplicate .js/.jsx context files into single .jsx files
- Removed 8 redundant .js context files (ETCswapContext, PriceContext,
  RoleContext, ThemeContext, UIContext, UserPreferencesContext,
  WalletContext, Web3Context)
- Updated contexts/index.js to import from consolidated .jsx files

**Unused Utilities**
- Removed metadataGenerator.js (only used in tests)
- Removed metadataGenerator.test.js

**Unused/Legacy Components**
- Removed TraditionalVoting.jsx/.css
- Removed GovernanceModeSelector.jsx/.css
- Removed MarketCreation.jsx/.css
- Removed WelfareMetrics.jsx
- Removed MyPositions.jsx/.css

**Demo Components and Mock Data**
- Removed StateManagementDemo.jsx/.css and /state-demo route
- Removed ComponentExamples.jsx/.css and /ui-components route
- Removed mock-data.json (51KB)
- Updated App.jsx to remove demo routes and imports
- Updated components/ui/index.js to remove ComponentExamples export

All changes have been verified to ensure no production code dependencies
were broken. Remaining components and utilities are actively used in the
application.
