--- frontend/src/components/MarketCreation.css
@@ -1,236 +1,0 @@
-/* MarketCreation Component Styles */
-
-.market-creation {
-  padding: 2rem;
-  max-width: 900px;
-  margin: 0 auto;
-}
-
-/* Header */
-.creation-header {
-  margin-bottom: 2rem;
-}
-
-.creation-header h2 {
-  font-size: 2rem;
-  margin-bottom: 0.75rem;
-  color: var(--text-primary);
-}
-
-.header-description {
-  font-size: 1.125rem;
-  color: var(--text-secondary);
-  line-height: 1.6;
-}
-
-/* Form */
-.creation-form {
-  background: var(--bg-light);
-  border: 2px solid var(--border-color);
-  border-radius: 12px;
-  padding: 2rem;
-}
-
-/* Form Sections */
-.form-section {
-  margin-bottom: 2.5rem;
-  padding-bottom: 2rem;
-  border-bottom: 1px solid rgba(45, 122, 79, 0.2);
-}
-
-.form-section:last-of-type {
-  border-bottom: none;
-}
-
-.form-section h3 {
-  font-size: 1.5rem;
-  margin-bottom: 1.5rem;
-  color: var(--text-primary);
-}
-
-/* Form Groups */
-.form-group {
-  margin-bottom: 1.5rem;
-}
-
-.form-group label {
-  display: block;
-  font-size: 0.95rem;
-  font-weight: 600;
-  color: var(--text-primary);
-  margin-bottom: 0.5rem;
-}
-
-.required {
-  color: var(--danger-color);
-  margin-left: 0.25rem;
-}
-
-.form-group input,
-.form-group textarea {
-  width: 100%;
-  padding: 0.75rem;
-  border: 1px solid var(--border-color);
-  border-radius: 8px;
-  background: var(--bg-dark);
-  color: var(--text-primary);
-  font-size: 1rem;
-  font-family: inherit;
-  transition: border-color 0.2s;
-}
-
-.form-group textarea {
-  resize: vertical;
-  min-height: 100px;
-}
-
-.form-group input:focus,
-.form-group textarea:focus {
-  outline: none;
-  border-color: var(--primary-color);
-  box-shadow: 0 0 0 3px rgba(45, 122, 79, 0.1);
-}
-
-.form-group input:focus-visible,
-.form-group textarea:focus-visible {
-  outline: 2px solid var(--primary-color);
-  outline-offset: 2px;
-}
-
-.form-group input[aria-invalid="true"],
-.form-group textarea[aria-invalid="true"] {
-  border-color: var(--danger-color);
-}
-
-.helper-text {
-  display: block;
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-  margin-top: 0.25rem;
-}
-
-.error-text {
-  display: block;
-  color: var(--danger-color);
-  font-size: 0.875rem;
-  margin-top: 0.25rem;
-  font-weight: 500;
-}
-
-/* Form Row */
-.form-row {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 1.5rem;
-}
-
-/* Info Notice */
-.info-notice {
-  display: flex;
-  gap: 0.75rem;
-  align-items: flex-start;
-  background: rgba(245, 158, 11, 0.1);
-  border: 1px solid var(--warning-color);
-  border-radius: 8px;
-  padding: 1.25rem;
-  margin-bottom: 2rem;
-}
-
-.info-notice > span:first-child {
-  font-size: 1.5rem;
-  flex-shrink: 0;
-  line-height: 1;
-}
-
-.notice-content {
-  flex: 1;
-  font-size: 0.95rem;
-  color: var(--text-primary);
-  line-height: 1.6;
-}
-
-.notice-content strong {
-  font-weight: 600;
-  color: var(--warning-color);
-}
-
-/* Form Actions */
-.form-actions {
-  display: flex;
-  justify-content: center;
-  padding-top: 1rem;
-}
-
-.create-button {
-  padding: 1rem 3rem;
-  background: linear-gradient(135deg, var(--primary-color) 0%, var(--secondary-color) 100%);
-  color: white;
-  border: none;
-  border-radius: 8px;
-  font-size: 1.125rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.3s;
-  min-width: 200px;
-}
-
-.create-button:hover:not(:disabled) {
-  transform: translateY(-2px);
-  box-shadow: 0 6px 16px rgba(45, 122, 79, 0.4);
-}
-
-.create-button:focus-visible {
-  outline: 2px solid var(--primary-color);
-  outline-offset: 4px;
-}
-
-.create-button:active:not(:disabled) {
-  transform: translateY(0);
-}
-
-.create-button:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-/* Responsive Design */
-@media (max-width: 768px) {
-  .market-creation {
-    padding: 1rem;
-  }
-
-  .creation-form {
-    padding: 1.5rem;
-  }
-
-  .creation-header h2 {
-    font-size: 1.5rem;
-  }
-
-  .header-description {
-    font-size: 1rem;
-  }
-
-  .form-section h3 {
-    font-size: 1.25rem;
-  }
-
-  .form-row {
-    grid-template-columns: 1fr;
-  }
-
-  .create-button {
-    width: 100%;
-  }
-}
-
-/* Reduced Motion Support */
-@media (prefers-reduced-motion: reduce) {
-  *,
-  *::before,
-  *::after {
-    animation-duration: 0.01ms !important;
-    animation-iteration-count: 1 !important;
-    transition-duration: 0.01ms !important;
-  }
-}
