--- frontend/src/components/MyPositions.css
@@ -1,408 +1,0 @@
-/* MyPositions Component Styles */
-
-/* Screen reader only content */
-.sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border-width: 0;
-}
-
-.my-positions {
-  padding: 2rem;
-  max-width: 1400px;
-  margin: 0 auto;
-}
-
-/* Header Section */
-.positions-header {
-  margin-bottom: 2rem;
-}
-
-.positions-header-top {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 1.5rem;
-  flex-wrap: wrap;
-  gap: 1rem;
-}
-
-.positions-header h2 {
-  font-size: 2rem;
-  margin: 0;
-  color: var(--text-primary);
-}
-
-/* Portfolio Summary */
-.portfolio-summary {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  gap: 1.5rem;
-  margin-bottom: 2rem;
-}
-
-.summary-card {
-  background: var(--bg-light);
-  border: 1px solid var(--border-color);
-  border-radius: 12px;
-  padding: 1.5rem;
-  text-align: center;
-}
-
-.summary-label {
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-  margin-bottom: 0.5rem;
-  font-weight: 600;
-}
-
-.summary-value {
-  font-size: 1.75rem;
-  font-weight: 700;
-  color: var(--text-primary);
-}
-
-.summary-value.profit {
-  color: var(--success-color);
-}
-
-.summary-value.loss {
-  color: var(--danger-color);
-}
-
-/* Filter Tabs */
-.filter-tabs {
-  display: flex;
-  gap: 1rem;
-  margin-bottom: 2rem;
-  border-bottom: 2px solid var(--border-color);
-}
-
-.filter-tab {
-  background: transparent;
-  border: none;
-  border-bottom: 3px solid transparent;
-  padding: 0.75rem 1.5rem;
-  color: var(--text-secondary);
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.3s;
-  margin-bottom: -2px;
-}
-
-.filter-tab:hover {
-  color: var(--text-primary);
-  background: rgba(45, 122, 79, 0.05);
-}
-
-.filter-tab:focus-visible {
-  outline: 2px solid var(--primary-color);
-  outline-offset: 2px;
-}
-
-.filter-tab.active {
-  color: var(--primary-color);
-  border-bottom-color: var(--primary-color);
-}
-
-/* Positions List */
-.positions-list {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-/* Position Card */
-.position-card {
-  background: var(--bg-light);
-  border: 2px solid var(--border-color);
-  border-radius: 12px;
-  padding: 1.5rem;
-  transition: all 0.3s ease;
-}
-
-.position-card:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 8px 20px rgba(45, 122, 79, 0.2);
-  border-color: var(--primary-color);
-}
-
-.position-card.settled {
-  opacity: 0.8;
-}
-
-/* Position Header */
-.position-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  margin-bottom: 1.5rem;
-  padding-bottom: 1rem;
-  border-bottom: 1px solid rgba(45, 122, 79, 0.2);
-}
-
-.position-title {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  flex: 1;
-}
-
-.position-title h3 {
-  font-size: 1.25rem;
-  color: var(--text-primary);
-  margin: 0;
-}
-
-/* Token Badge */
-.token-badge {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.25rem;
-  padding: 0.25rem 0.75rem;
-  border-radius: 20px;
-  font-size: 0.875rem;
-  font-weight: 600;
-}
-
-.token-badge.pass {
-  background: rgba(34, 197, 94, 0.15);
-  color: var(--success-color);
-  border: 1px solid var(--success-color);
-}
-
-.token-badge.fail {
-  background: rgba(220, 38, 38, 0.15);
-  color: var(--danger-color);
-  border: 1px solid var(--danger-color);
-}
-
-/* Status Badge */
-.status-badge {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.25rem;
-  padding: 0.25rem 0.75rem;
-  border-radius: 20px;
-  font-size: 0.875rem;
-  font-weight: 600;
-}
-
-.status-badge.active {
-  background: var(--success-color);
-  color: white;
-}
-
-.status-badge.active::before {
-  content: '✓';
-}
-
-.status-badge.settled {
-  background: var(--text-secondary);
-  color: white;
-}
-
-.status-badge.settled::before {
-  content: '✓';
-}
-
-/* Position Details */
-.position-details {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  margin-bottom: 1rem;
-}
-
-.detail-row {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
-  gap: 1rem;
-}
-
-.detail-item {
-  display: flex;
-  flex-direction: column;
-  gap: 0.25rem;
-}
-
-.detail-label {
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-}
-
-.detail-value {
-  font-size: 1rem;
-  color: var(--text-primary);
-  font-weight: 500;
-}
-
-.detail-value.pnl {
-  font-weight: 700;
-  font-size: 1.125rem;
-  display: flex;
-  align-items: center;
-  gap: 0.25rem;
-}
-
-.detail-value.pnl.profit {
-  color: var(--success-color);
-}
-
-.detail-value.pnl.loss {
-  color: var(--danger-color);
-}
-
-/* Settlement Info */
-.settlement-info {
-  padding: 0.75rem;
-  background: rgba(148, 163, 184, 0.1);
-  border-radius: 8px;
-  font-size: 0.9rem;
-  color: var(--text-secondary);
-  margin-top: 0.5rem;
-}
-
-/* Position Actions */
-.position-actions {
-  display: flex;
-  gap: 1rem;
-  padding-top: 1rem;
-  border-top: 1px solid rgba(45, 122, 79, 0.2);
-}
-
-.exit-button {
-  flex: 1;
-  padding: 0.75rem 1.5rem;
-  background: linear-gradient(135deg, var(--primary-color) 0%, var(--secondary-color) 100%);
-  color: white;
-  border: none;
-  border-radius: 8px;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.3s;
-}
-
-.exit-button:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(45, 122, 79, 0.4);
-}
-
-.exit-button:focus-visible {
-  outline: 2px solid var(--primary-color);
-  outline-offset: 4px;
-}
-
-.exit-button:active {
-  transform: translateY(0);
-}
-
-/* Loading State */
-.loading {
-  text-align: center;
-  padding: 3rem;
-  color: var(--text-secondary);
-  font-size: 1.125rem;
-}
-
-.loading::before {
-  content: '';
-  display: block;
-  width: 40px;
-  height: 40px;
-  margin: 0 auto 1rem;
-  border: 3px solid rgba(45, 122, 79, 0.2);
-  border-top-color: var(--primary-color);
-  border-radius: 50%;
-  animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-  to { transform: rotate(360deg); }
-}
-
-/* Empty States */
-.no-positions,
-.empty-filter {
-  text-align: center;
-  padding: 3rem;
-  color: var(--text-secondary);
-}
-
-.no-positions .placeholder-icon,
-.empty-filter .placeholder-icon {
-  font-size: 4rem;
-  margin-bottom: 1rem;
-}
-
-.no-positions p,
-.empty-filter p {
-  font-size: 1.125rem;
-  line-height: 1.6;
-}
-
-/* Responsive Design */
-@media (max-width: 768px) {
-  .my-positions {
-    padding: 1rem;
-  }
-
-  .positions-header h2 {
-    font-size: 1.5rem;
-  }
-
-  .portfolio-summary {
-    grid-template-columns: 1fr;
-  }
-
-  .filter-tabs {
-    flex-wrap: wrap;
-  }
-
-  .filter-tab {
-    padding: 0.5rem 1rem;
-    font-size: 0.9rem;
-  }
-
-  .position-header {
-    flex-direction: column;
-    gap: 1rem;
-  }
-
-  .position-title {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .detail-row {
-    grid-template-columns: 1fr;
-  }
-
-  .position-actions {
-    flex-direction: column;
-  }
-}
-
-/* Reduced Motion Support */
-@media (prefers-reduced-motion: reduce) {
-  *,
-  *::before,
-  *::after {
-    animation-duration: 0.01ms !important;
-    animation-iteration-count: 1 !important;
-    transition-duration: 0.01ms !important;
-  }
-}
