--- frontend/src/components/StateManagementDemo.css
@@ -1,354 +1,0 @@
-.demo-container {
-  max-width: 1400px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
-.demo-header {
-  text-align: center;
-  margin-bottom: 3rem;
-}
-
-.demo-header h1 {
-  font-size: 2.5rem;
-  margin-bottom: 0.5rem;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  background-clip: text;
-}
-
-.demo-header p {
-  color: #6b7280;
-  font-size: 1.1rem;
-}
-
-.demo-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(400px, 1fr));
-  gap: 2rem;
-}
-
-.demo-card {
-  background: white;
-  border-radius: 12px;
-  padding: 2rem;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-  border: 1px solid #e5e7eb;
-}
-
-.demo-card.full-width {
-  grid-column: 1 / -1;
-}
-
-.demo-card.info-card {
-  background: linear-gradient(135deg, #667eea15 0%, #764ba215 100%);
-}
-
-.demo-card h2 {
-  margin: 0 0 1.5rem 0;
-  font-size: 1.5rem;
-  color: #111827;
-}
-
-.state-info {
-  margin-bottom: 1.5rem;
-}
-
-.state-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.75rem 0;
-  border-bottom: 1px solid #e5e7eb;
-}
-
-.state-item:last-child {
-  border-bottom: none;
-}
-
-.state-label {
-  font-weight: 600;
-  color: #374151;
-}
-
-.state-value {
-  font-size: 0.95rem;
-  color: #6b7280;
-}
-
-.state-value.monospace {
-  font-family: 'Courier New', monospace;
-  font-size: 0.85rem;
-}
-
-.state-value.connected {
-  color: #10b981;
-  font-weight: 600;
-}
-
-.state-value.disconnected {
-  color: #ef4444;
-  font-weight: 600;
-}
-
-.demo-actions,
-.demo-actions-grid {
-  display: flex;
-  gap: 1rem;
-  flex-wrap: wrap;
-}
-
-.demo-actions-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
-}
-
-.primary-button,
-.secondary-button,
-.action-button {
-  padding: 0.75rem 1.5rem;
-  border: none;
-  border-radius: 8px;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.primary-button {
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-}
-
-.primary-button:hover:not(:disabled) {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(102, 126, 234, 0.4);
-}
-
-.secondary-button {
-  background: white;
-  color: #667eea;
-  border: 2px solid #667eea;
-}
-
-.secondary-button:hover:not(:disabled) {
-  background: #f3f4f6;
-}
-
-.action-button {
-  background: #f9fafb;
-  color: #374151;
-  border: 1px solid #e5e7eb;
-}
-
-.action-button:hover:not(:disabled) {
-  background: #f3f4f6;
-  border-color: #d1d5db;
-}
-
-.primary-button:disabled,
-.secondary-button:disabled,
-.action-button:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-  transform: none !important;
-}
-
-.error-display {
-  margin-top: 1.5rem;
-  padding: 1rem;
-  background: #fef2f2;
-  border: 1px solid #fecaca;
-  border-radius: 8px;
-}
-
-.error-display h3 {
-  margin: 0 0 0.5rem 0;
-  color: #991b1b;
-  font-size: 1rem;
-}
-
-.error-display p {
-  margin: 0 0 1rem 0;
-  color: #7f1d1d;
-}
-
-.event-log {
-  background: #f9fafb;
-  border: 1px solid #e5e7eb;
-  border-radius: 8px;
-  padding: 1rem;
-  max-height: 300px;
-  overflow-y: auto;
-  margin-bottom: 1rem;
-}
-
-.empty-log {
-  text-align: center;
-  color: #6b7280;
-  padding: 2rem;
-  margin: 0;
-}
-
-.log-entry {
-  display: flex;
-  gap: 1rem;
-  padding: 0.5rem;
-  border-bottom: 1px solid #e5e7eb;
-  animation: slideIn 0.3s ease-out;
-}
-
-.log-entry:last-child {
-  border-bottom: none;
-}
-
-@keyframes slideIn {
-  from {
-    opacity: 0;
-    transform: translateX(-10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateX(0);
-  }
-}
-
-.log-time {
-  font-family: 'Courier New', monospace;
-  font-size: 0.85rem;
-  color: #6b7280;
-  min-width: 80px;
-}
-
-.log-message {
-  color: #374151;
-  flex: 1;
-}
-
-.info-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  gap: 1.5rem;
-}
-
-.info-item h3 {
-  margin: 0 0 0.5rem 0;
-  color: #667eea;
-  font-size: 1.1rem;
-}
-
-.info-item p {
-  margin: 0 0 0.75rem 0;
-  color: #6b7280;
-  font-size: 0.95rem;
-  line-height: 1.5;
-}
-
-.info-item code {
-  display: block;
-  background: white;
-  padding: 0.5rem;
-  border-radius: 4px;
-  font-size: 0.85rem;
-  color: #667eea;
-  font-family: 'Courier New', monospace;
-  border: 1px solid #e5e7eb;
-}
-
-.network-error-banner {
-  position: fixed;
-  bottom: 20px;
-  left: 50%;
-  transform: translateX(-50%);
-  background: #fef2f2;
-  color: #991b1b;
-  padding: 1rem 2rem;
-  border-radius: 8px;
-  border: 2px solid #fecaca;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-  max-width: 600px;
-  text-align: center;
-  animation: slideUp 0.3s ease-out;
-  z-index: 1000;
-}
-
-@keyframes slideUp {
-  from {
-    opacity: 0;
-    transform: translateX(-50%) translateY(20px);
-  }
-  to {
-    opacity: 1;
-    transform: translateX(-50%) translateY(0);
-  }
-}
-
-.demo-modal-content {
-  padding: 1rem 0;
-}
-
-.demo-modal-content p {
-  margin: 0 0 1rem 0;
-  color: #374151;
-}
-
-.modal-actions {
-  display: flex;
-  gap: 1rem;
-  margin-top: 1.5rem;
-}
-
-/* Responsive */
-@media (max-width: 768px) {
-  .demo-container {
-    padding: 1rem;
-  }
-
-  .demo-header h1 {
-    font-size: 2rem;
-  }
-
-  .demo-grid {
-    grid-template-columns: 1fr;
-    gap: 1rem;
-  }
-
-  .demo-card {
-    padding: 1.5rem;
-  }
-
-  .demo-actions,
-  .demo-actions-grid {
-    flex-direction: column;
-  }
-
-  .action-button,
-  .primary-button,
-  .secondary-button {
-    width: 100%;
-  }
-
-  .info-grid {
-    grid-template-columns: 1fr;
-  }
-
-  .network-error-banner {
-    left: 10px;
-    right: 10px;
-    transform: none;
-    max-width: none;
-  }
-}
-
-/* Reduced motion */
-@media (prefers-reduced-motion: reduce) {
-  .log-entry,
-  .network-error-banner {
-    animation: none;
-  }
-  
-  .primary-button:hover:not(:disabled) {
-    transform: none;
-  }
-}
