--- frontend/src/components/ui/ComponentExamples.css
@@ -1,143 +1,0 @@
-/* Component Examples Styles */
-
-.component-examples {
-  min-height: 100vh;
-  background: #0f1810;
-  color: #f1f5f9;
-  padding: 2rem;
-}
-
-.examples-container {
-  max-width: 1200px;
-  margin: 0 auto;
-}
-
-.component-examples h1 {
-  font-size: 2.5rem;
-  color: #2D7A4F;
-  margin-bottom: 0.5rem;
-}
-
-.intro {
-  font-size: 1.125rem;
-  color: #94a3b8;
-  margin-bottom: 3rem;
-}
-
-.example-section {
-  margin-bottom: 4rem;
-}
-
-.example-section h2 {
-  font-size: 2rem;
-  color: #34A853;
-  margin-bottom: 1.5rem;
-  border-bottom: 2px solid #2D7A4F;
-  padding-bottom: 0.5rem;
-}
-
-.example-section h3 {
-  font-size: 1.25rem;
-  color: #f1f5f9;
-  margin-bottom: 1rem;
-}
-
-.example-section h4 {
-  font-size: 1.125rem;
-  color: #f1f5f9;
-  margin-bottom: 0.5rem;
-}
-
-.example-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-  gap: 2rem;
-  margin-bottom: 2rem;
-}
-
-.example-item {
-  margin-bottom: 1.5rem;
-}
-
-.example-badges,
-.example-statuses {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  margin-bottom: 2rem;
-}
-
-.form-actions {
-  display: flex;
-  gap: 1rem;
-  margin-top: 2rem;
-}
-
-.accessibility-list {
-  list-style: none;
-  padding: 0;
-  margin: 1rem 0;
-}
-
-.accessibility-list li {
-  padding: 0.5rem 0;
-  font-size: 1rem;
-  color: #94a3b8;
-}
-
-.usage-code {
-  background: #1a2820;
-  border: 1px solid #2D7A4F;
-  border-radius: 8px;
-  padding: 1.5rem;
-  overflow-x: auto;
-  margin-top: 1rem;
-}
-
-.usage-code pre {
-  margin: 0;
-  color: #f1f5f9;
-  font-family: 'Courier New', monospace;
-  font-size: 0.875rem;
-  line-height: 1.6;
-}
-
-/* Responsive */
-@media (max-width: 768px) {
-  .component-examples {
-    padding: 1rem;
-  }
-
-  .component-examples h1 {
-    font-size: 2rem;
-  }
-
-  .example-section h2 {
-    font-size: 1.5rem;
-  }
-
-  .example-grid {
-    grid-template-columns: 1fr;
-  }
-
-  .form-actions {
-    flex-direction: column;
-  }
-
-  .usage-code {
-    padding: 1rem;
-  }
-
-  .usage-code pre {
-    font-size: 0.75rem;
-  }
-}
-
-/* Reduced Motion Support */
-@media (prefers-reduced-motion: reduce) {
-  * {
-    animation-duration: 0.01ms !important;
-    animation-iteration-count: 1 !important;
-    transition-duration: 0.01ms !important;
-  }
-}
