+++ frontend/src/components/ui/index.js
@@ -12,7 +12,6 @@ export { default as StatusIndicator } from './StatusIndicator'
 export { default as Input } from './Input'
 export { default as FormGroup } from './FormGroup'
 export { default as HelperText } from './HelperText'
-export { default as ComponentExamples } from './ComponentExamples'
 export { default as ThemeToggle } from './ThemeToggle'
 export { default as UserManagementModal } from './UserManagementModal'
 export { default as FairWinsUserModal } from './FairWinsUserModal'
