--- frontend/src/contexts/ETCswapContext.js
@@ -1,3 +1,0 @@
-import { createContext } from 'react'
-
-export const ETCswapContext = createContext(null)
