+++ frontend/src/contexts/PriceContext.jsx
@@ -1,5 +1,15 @@
+import { createContext, useContext } from 'react'
 import usePriceConversion from '../hooks/usePriceConversion'
-import PriceContext from './PriceContext'
+
+const PriceContext = createContext(null)
+
+export function usePrice() {
+  const context = useContext(PriceContext)
+  if (!context) {
+    throw new Error('usePrice must be used within a PriceProvider')
+  }
+  return context
+}
 
 export function PriceProvider({ children }) {
   const priceData = usePriceConversion()
@@ -10,3 +20,5 @@ export function PriceProvider({ children }) {
     </PriceContext.Provider>
   )
 }
+
+export default PriceContext
