+++ frontend/src/contexts/RoleContext.jsx
@@ -1,12 +1,49 @@
-import { useState, useEffect, useCallback } from 'react'
+import { createContext, useState, useEffect, useCallback } from 'react'
 import { useWeb3 } from '../hooks/useWeb3'
-import { 
-  getUserRoles, 
+import {
+  getUserRoles,
   hasRole as checkRole,
   addUserRole,
   removeUserRole
 } from '../utils/roleStorage'
-import { RoleContext, ROLES, ROLE_INFO } from './RoleContext'
+
+export const RoleContext = createContext(null)
+
+/**
+ * Available roles in the system
+ */
+export const ROLES = {
+  MARKET_MAKER: 'MARKET_MAKER',
+  CLEARPATH_USER: 'CLEARPATH_USER',
+  TOKENMINT: 'TOKENMINT',
+  ADMIN: 'ADMIN', // For role management
+}
+
+/**
+ * Role display names and descriptions
+ */
+export const ROLE_INFO = {
+  [ROLES.MARKET_MAKER]: {
+    name: 'Market Maker',
+    description: 'Ability to create and manage prediction markets',
+    premium: true
+  },
+  [ROLES.CLEARPATH_USER]: {
+    name: 'ClearPath User',
+    description: 'Access to DAO governance and management platform',
+    premium: true
+  },
+  [ROLES.TOKENMINT]: {
+    name: 'Token Mint',
+    description: 'Access to mint and manage NFTs and ERC20 tokens',
+    premium: true
+  },
+  [ROLES.ADMIN]: {
+    name: 'Administrator',
+    description: 'Full system access including role management',
+    premium: false
+  }
+}
 
 /**
  * RoleProvider manages user roles tied to wallet address
