--- frontend/src/contexts/ThemeContext.js
@@ -1,3 +1,0 @@
-import { createContext } from 'react'
-
-export const ThemeContext = createContext(null)
