--- frontend/src/contexts/UIContext.js
@@ -1,3 +1,0 @@
-import { createContext } from 'react'
-
-export const UIContext = createContext(null)
