--- frontend/src/contexts/UserPreferencesContext.js
@@ -1,3 +1,0 @@
-import { createContext } from 'react'
-
-export const UserPreferencesContext = createContext(null)
