--- frontend/src/contexts/Web3Context.js
@@ -1,3 +1,0 @@
-import { createContext } from 'react'
-
-export const Web3Context = createContext(null)
