+++ frontend/src/contexts/index.js
@@ -1,17 +1,8 @@
-export { Web3Context } from './Web3Context.js'
-export { Web3Provider } from './Web3Context.jsx'
-export { UIContext } from './UIContext.js'
-export { UIProvider } from './UIContext.jsx'
-export { ThemeContext } from './ThemeContext.js'
-export { ThemeProvider } from './ThemeContext.jsx'
-export { usePrice } from './PriceContext.js'
-export { PriceProvider } from './PriceContext.jsx'
-export { ETCswapContext } from './ETCswapContext.js'
-export { ETCswapProvider } from './ETCswapContext.jsx'
-export { UserPreferencesContext } from './UserPreferencesContext.js'
-export { UserPreferencesProvider } from './UserPreferencesContext.jsx'
-export { RoleContext, ROLES, ROLE_INFO } from './RoleContext.js'
-export { RoleProvider } from './RoleContext.jsx'
-export { WalletContext } from './WalletContext.js'
-export { WalletProvider } from './WalletContext.jsx'
-
+export { Web3Context, Web3Provider } from './Web3Context.jsx'
+export { UIContext, UIProvider } from './UIContext.jsx'
+export { ThemeContext, ThemeProvider } from './ThemeContext.jsx'
+export { usePrice, PriceProvider } from './PriceContext.jsx'
+export { ETCswapContext, ETCswapProvider } from './ETCswapContext.jsx'
+export { UserPreferencesContext, UserPreferencesProvider } from './UserPreferencesContext.jsx'
+export { RoleContext, ROLES, ROLE_INFO, RoleProvider } from './RoleContext.jsx'
+export { WalletContext, WalletProvider } from './WalletContext.jsx'
